pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trello-mcp"
version = "0.1.0"
description = "MCP server for Trello - manage boards, lists, and cards"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.4.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
]

[project.scripts]
trello-mcp = "trello_mcp.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
